Инициализация и работа с интерфейсом из Render вынесены в класс GUI
