+++ src/main/java/tow/engine/cycle/Engine.java
@@ -10,30 +10,28 @@ public class Engine{
 
 	public Update update;
 	public Render render;
+	public GUI gui;
 	public Analyzer analyzer;
 	private FPSLimit fpsLimit;
-	
-	public Engine(){
+
+	public void init(){
 		update = new Update();
 		render = new Render();
+		gui = new GUI();
 		analyzer = new Analyzer();
 		fpsLimit = new FPSLimit(SettingsStorage.DISPLAY.FPS_LIMIT);
 	}
 	
 	public void run(){
 		while(!GLFW.glfwWindowShouldClose(render.getWindowID())){
-			//Получение событий и Callbacks
-
 			//Цикл Update
 			analyzer.startUpdate();
-			render.pollEvents();
-			update.loop();
-			analyzer.update();
+			update.loop(); //Обновление состояния у всех объектов в активной локации
 			analyzer.endUpdate();
 
 			//Цикл Render
 			analyzer.startRender();
-			render.loop();
+			render.loop(); //Отрисовка всех объектов в активной локациии
 			analyzer.endRender();
 
 			//Пауза для синхронизации кадров
